Add patch unit test

+++ models.py
@@ -12,7 +12,7 @@ class Cupcake(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     flavor = db.Column(db.Text, nullable=False)
     size = db.Column(db.Text, nullable=False)
-    rating = db.Column(db.Float, nullable=False)
+    rating = db.Column(db.Integer, nullable=False)
     image = db.Column(db.Text, nullable=False, default="https://tinyurl.com/demo-cupcake")
 
     # Returns dictionary representation of cupcakes to jsonify
